requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
asn1crypto==1.5.1
attrs==24.3.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.16.0
colorama==0.4.6
cryptography==44.0.0
DateTime==5.5
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
google-auth==2.37.0
greenlet==3.1.1
gunicorn==23.0.0
html2text==2024.2.26
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
pg8000==1.31.2
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
scramp==1.4.5
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
zope.interface==7.2
